Add serial debug logging
For reasons way above my head "%d" does not work in the format string
for kprintf?! It keeps printing 0 when used. But, when I use snprintf to
first generate the string and then log it using kprintf all is well.

Wont pursue this further for now even though it creates a small overhead
when logging.